backend-fastapi/data-access-service/requirements.txt
fastapi
uvicorn
langchain
langchain-postgres
python-dotenv
psycopg[binary]
requests
